--- src/components/AddActivityBtn.tsx
@@ -1,9 +1,0 @@
-import { Button } from '@mui/material';
-
-
-const AddActivityBtn: React.FunctionComponent = () => (
-    <Button variant='contained'>Legg til ny aktivitet</Button>
-);
-
-export default AddActivityBtn;
-
