--- src/components/JoinedActivitiesBtn.tsx
@@ -1,9 +1,0 @@
-import { Button } from '@mui/material';
-
-
-const JoinedActivitiesBtn = () => (
-    <Button variant='contained'>Påmeldte aktiviteter</Button>
-);
-
-export default JoinedActivitiesBtn;
-
