--- src/components/MyActivitiesBtn.tsx
@@ -1,9 +1,0 @@
-import { Button } from '@mui/material';
-
-
-const MyActivitiesBtn = () => (
-    <Button variant='contained'>Mine aktiviteter</Button>
-);
-
-export default MyActivitiesBtn;
-
